音频引擎适配时延查询功能

+++ services/audio_engine/manager/src/hpae_offload_renderer_manager.cpp
@@ -748,7 +748,7 @@ void HpaeOffloadRendererManager::OnNodeStatusUpdate(uint32_t sessionId, IOperati
 
 void HpaeOffloadRendererManager::OnRequestLatency(uint32_t sessionId, uint64_t &latency)
 {
-    uint32_t processLatency = 0;
+    uint64_t processLatency = 0;
 
     if (converterForLoudness_) {
         processLatency += converterForLoudness_->GetLatency();
